ref: Move colibri v1 related code to a v1 package.
