+++ src/main/java/org/jitsi/jicofo/conference/colibri/v1/BridgeSession.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jitsi.jicofo.conference.colibri;
+package org.jitsi.jicofo.conference.colibri.v1;
 
 import edu.umd.cs.findbugs.annotations.*;
 import org.jetbrains.annotations.*;
@@ -23,6 +23,7 @@ import org.jitsi.impl.protocol.xmpp.*;
 import org.jitsi.jicofo.*;
 import org.jitsi.jicofo.bridge.*;
 import org.jitsi.jicofo.conference.*;
+import org.jitsi.jicofo.conference.colibri.*;
 import org.jitsi.jicofo.conference.source.*;
 import org.jitsi.utils.logging2.*;
 import org.jitsi.xmpp.extensions.colibri.*;
