+++ src/main/java/org/jitsi/jicofo/conference/colibri/v1/OctoParticipant.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jitsi.jicofo.conference.colibri;
+package org.jitsi.jicofo.conference.colibri.v1;
 
 import org.jitsi.jicofo.conference.*;
 import org.jitsi.jicofo.conference.source.*;
