+++ src/main/kotlin/org/jitsi/jicofo/JicofoServices.kt
@@ -32,7 +32,7 @@ import org.jitsi.jicofo.auth.XMPPDomainAuthAuthority
 import org.jitsi.jicofo.bridge.BridgeConfig
 import org.jitsi.jicofo.bridge.BridgeMucDetector
 import org.jitsi.jicofo.bridge.BridgeSelector
-import org.jitsi.jicofo.conference.colibri.ColibriConferenceImpl
+import org.jitsi.jicofo.conference.colibri.v1.ColibriConferenceImpl
 import org.jitsi.jicofo.health.HealthConfig
 import org.jitsi.jicofo.health.JicofoHealthChecker
 import org.jitsi.jicofo.jibri.JibriConfig
