+++ src/main/kotlin/org/jitsi/jicofo/conference/colibri/v1/ParticipantInfo.kt
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jitsi.jicofo.conference.colibri
+package org.jitsi.jicofo.conference.colibri.v1
 
 import org.jitsi.xmpp.extensions.colibri.ColibriConferenceIQ
 import org.jitsi.xmpp.extensions.jingle.IceUdpTransportPacketExtension
