+++ src/test/java/org/jitsi/jicofo/ColibriThreadingTest.java
@@ -24,7 +24,7 @@ import mock.jvb.*;
 import mock.xmpp.*;
 import mock.xmpp.colibri.*;
 import org.jitsi.jicofo.codec.*;
-import org.jitsi.jicofo.conference.colibri.*;
+import org.jitsi.jicofo.conference.colibri.v1.*;
 import org.jitsi.xmpp.extensions.jingle.*;
 import org.junit.jupiter.api.*;
 import org.jxmpp.jid.*;
